Moved Auth methods to Auth class
